refactor(settings): rozdziel test od local, usuń martwy docker.py
Moduł settings miał dwa pliki udające żywą konfigurację, których nic
nie ładowało:

- `test.py` (dawny `master.py`) — testy de facto jechały na `local.py`
  (pytest.ini --ds=settings.local), a `local.py` pełnił podwójną rolę
  dev+test przez blok `if "pytest" in sys.modules`.
- `docker.py` — nieużywany (.env.docker → settings.local, Dockerfile →
  settings.production).

Zmiany:

- `local.py` → czysto deweloperski; usunięty blok pytest (+ martwy
  `import sys`).
- `test.py` → ożywiony jako cienka warstwa `from .local import *` +
  test-only tweaki (usunięcie cacheops, CACHEOPS_ENABLED=False, drop
  first_run_wizard middleware) nałożone bezwarunkowo.
- `pytest.ini` → --ds=django_bpp.settings.test (jawny łańcuch
  test → local → base).
- `docker.py` usunięty.
- `wsgi.py`/`asgi.py` → naprawiony fallback DJANGO_SETTINGS_MODULE
  z pustego pakietu `django_bpp.settings` na `...settings.local`
  (spójnie z manage.py; w produkcji ENV i tak ustawia production).
- dwa pliki testowe → fallback literal local → test dla spójności.

Zachowanie testów i deva niezmienione (zweryfikowane wartość-po-wartości
przez django.setup() oraz 17 testów najczulszych na settings).
