+++ pytest.ini
@@ -19,7 +19,7 @@ addopts =
     --durations=50 --durations-min=1.0
     --maxfail 50
     --timeout 90
-    --ds=django_bpp.settings.local
+    --ds=django_bpp.settings.test
     --only-rerun TimeoutError
     --only-rerun TimeoutException
     --only-rerun ElementClickInterceptedException
