+++ src/django_bpp/asgi.py
@@ -3,7 +3,7 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_bpp.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_bpp.settings.local")
 # Initialize Django ASGI application early to ensure the AppRegistry
 # is populated before importing code that may import ORM models.
 django_asgi_app = get_asgi_application()
